Typo in ExtendedRandom, improved BackgroundBuilder.withColor() and added a changeable time scale to Scene

+++ gdx/src/com/upseil/gdx/math/ExtendedRandom.java
@@ -74,7 +74,7 @@ public interface ExtendedRandom {
     /**
      * Returns a random number between start (inclusive) and end (exclusive).
      */
-    public default double randomFloat(double start, double end) {
+    public default double randomDouble(double start, double end) {
         return start + randomDouble() * (end - start);
     }
